refactor trace store manifest node attach
